+++ src/Pasargad.php
@@ -3,6 +3,8 @@
 namespace Pasargad;
 
 use Pasargad\Classes\AbstractPayment;
+use Pasargad\Classes\PaymentItem;
+use Pasargad\Classes\PaymentType;
 use Pasargad\Classes\RequestBuilder;
 use Pasargad\Classes\RSA\RSAProcessor;
 
@@ -14,12 +16,21 @@ class Pasargad extends AbstractPayment
     /** @var string $token */
     private $token = null;
 
+    /** @var boolean $safeMode */
+    private $safeMode = true;
+
+    /** @var array $paymentItems */
+    private $paymentItems = [];
+
+    /** @var boolean $multiPaymentMode */
+    private $multiPaymentMode = false;
+
     /**
      * Address of gateway for getting token
      * @var string
      */
     const URL_GET_TOKEN = "https://pep.shaparak.ir/Api/v1/Payment/GetToken";
- 
+
     /**
      * Redirect User with token to this URL
      * e.q: https://pep.shaparak.ir/payment.aspx?n=Token
@@ -37,13 +48,15 @@ class Pasargad extends AbstractPayment
      * @var string $certificateFile
      * @var string $action
      */
-    public function __construct($merchantCode, $terminalCode, $redirectAddress, $certificateFile, $action = "1003")
+    public function __construct($merchantCode, $terminalCode, $redirectAddress, $certificateFile, $merchantName = null, $action = "1003", $safeMode = true)
     {
         $this->merchantId = $merchantCode;
         $this->terminalId = $terminalCode;
         $this->redirectUrl = $redirectAddress;
         $this->certificate = $certificateFile;
+        $this->merchantName = $merchantName;
         $this->action = $action;
+        $this->safeMode = $safeMode;
         $this->api = new RequestBuilder();
     }
 
@@ -70,8 +83,22 @@ class Pasargad extends AbstractPayment
         $params['terminalCode'] = $this->getTerminalId();
         $params['redirectAddress'] = $this->getRedirectUrl();
         $params['timeStamp'] = date("Y/m/d H:i:s");
+        if ($this->getMobile()) {
+            $params['mobile'] = $this->getMobile();
+        }
+        if ($this->getEmail()) {
+            $params['email'] = $this->getEmail();
+        }
+        
+        if ($this->multiPaymentMode) { 
+            if ($this->getMerchantName() !== null) {
+                $params['MerchantName'] = $this->getMerchantName();
+            }
+            $params['MultiPaymentData'] = base64_encode($this->generatePayment());
+        }
+
         $sign = $this->sign(json_encode($params));
-        $this->token = $this->api->send(static::URL_GET_TOKEN, RequestBuilder::POST, ["Sign" => $sign], $params, true);
+        $this->token = $this->api->send(static::URL_GET_TOKEN, RequestBuilder::POST, ["Sign" => $sign], $params, true, $this->safeMode);
         return $this->token;
     }
 
@@ -100,7 +127,7 @@ class Pasargad extends AbstractPayment
         $params['terminalCode'] = $this->getTerminalId();
         $params['timeStamp'] = date("Y/m/d H:i:s");
         $sign = $this->sign(json_encode($params));
-        $response = $this->api->send(static::URL_VERIFY_PAYMENT,RequestBuilder::POST,["Sign" => $sign],$params, true);
+        $response = $this->api->send(static::URL_VERIFY_PAYMENT, RequestBuilder::POST, ["Sign" => $sign], $params, true, $this->safeMode);
         return $response;
     }
 
@@ -114,7 +141,7 @@ class Pasargad extends AbstractPayment
         $params['merchantCode'] = $this->getMerchantId();
         $params['terminalCode'] = $this->getTerminalId();
         $params['transactionReferenceID'] = $this->getTransactionReferenceId();
-        $response = $this->api->send(static::URL_CHECK_TRANSACTION,RequestBuilder::POST,[],$params, true);
+        $response = $this->api->send(static::URL_CHECK_TRANSACTION, RequestBuilder::POST, [], $params, true, $this->safeMode);
         return $response;
     }
 
@@ -129,7 +156,52 @@ class Pasargad extends AbstractPayment
         $params['terminalCode'] = $this->getTerminalId();
         $params['timeStamp'] = date("Y/m/d H:i:s");
         $sign = $this->sign(json_encode($params));
-        $response = $this->api->send(static::URL_VERIFY_PAYMENT,RequestBuilder::POST,["Sign" => $sign],$params, true);
+        $response = $this->api->send(static::URL_REFUND, RequestBuilder::POST, ["Sign" => $sign], $params, true, $this->safeMode);
         return $response;
+    }
+
+    public function addPaymentType($iban,$type,$value) 
+    {  
+        $this->paymentItems[] = new PaymentItem($iban,$type,$value);
+        $this->multiPaymentMode = true;
+    }
+
+
+    private function generatePayment()
+    {
+        $xw = xmlwriter_open_memory();
+        xmlwriter_set_indent($xw, 1);
+        $res = xmlwriter_set_indent_string($xw, ' ');
+        xmlwriter_start_document($xw, '1.0', 'UTF-8');
+        /** @var PaymentItem $item */
+        foreach ($this->paymentItems as $item) {
+            // <item>
+            xmlwriter_start_element($xw, 'item');
+
+                // <iban>
+                xmlwriter_start_element($xw, 'iban');
+                xmlwriter_text($xw, $item->getIban());
+                xmlwriter_end_element($xw); 
+                // </iban>
+
+
+                // <type>
+                xmlwriter_start_element($xw, 'type');
+                xmlwriter_text($xw, $item->getType());
+                xmlwriter_end_element($xw); 
+                // </type>
+
+                // <value>
+                xmlwriter_start_element($xw, 'value');
+                xmlwriter_text($xw, $item->getValue());
+                xmlwriter_end_element($xw); 
+                // </value>
+
+            xmlwriter_end_element($xw); 
+            // </item>
+        }
+
+        xmlwriter_end_document($xw);
+        return xmlwriter_output_memory($xw);
     }
 }