+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 All notable changes to this project is documented in this file.
+## [0.4.0] - 
+
+### Added
+ - MultiPayment (Tas-him) Implementation
+ - BossIPG (Basket Of ServiceS) Implementation
+
+### Changes
+ - Add `Mobile` and `Email` fields to `getToken()` method
+ - Add SafeMode to check or not checking ssl certificate in curl
+ - debug refund method
+
+
 ## [0.3.1] - 2021 15 Sept (namespace conflicts)
 ### Changes
  - namespace typo fixed